changed types for params Cell

+++ src/domain/RenderServiceImpl.kt
@@ -50,8 +50,8 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
             val row = floor(Math.random() * config.size).toInt()
             val coll = floor(Math.random() * config.size).toInt()
 
-            if (cellList[row][coll].value == 0) {
-                cellList[row][coll].value = 2 * ceil(Math.random() * 2).toInt()
+            if (cellList[row][coll].value == 0.0) {
+                cellList[row][coll].value = 2 * ceil(Math.random() * 2)
                 drawAllCells()
                 return
             }
@@ -65,14 +65,14 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
                 { innerCycle -> (innerCycle - 1) >= 0 }) { innerCycle, externalCycle ->
             val shiftCell = cellList[externalCycle][innerCycle - 1]
             val currentCell = cellList[externalCycle][innerCycle]
-            if (shiftCell.value == 0) {
+            if (shiftCell.value == 0.0) {
                 shiftCell.value = currentCell.value
-                currentCell.value = 0
+                currentCell.value = 0.0
                 return@moveUpLeftTemplate ActionMove.EMPTY_MOVE
             } else if (currentCell.value == shiftCell.value) {
                 shiftCell.value *= 2
-                scoreObservable.setValue(scoreObservable.getValue()!! + shiftCell.value)
-                currentCell.value = 0
+                scoreObservable.setValue((scoreObservable.getValue()!! + shiftCell.value).toInt())
+                currentCell.value = 0.0
                 return@moveUpLeftTemplate ActionMove.SUCCESS_MOVE
             } else {
                 return@moveUpLeftTemplate ActionMove.FAILED_MOVE
@@ -86,14 +86,14 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
             val shiftCell = cellList[innerCycle - 1][externalCycle]
             val currentCell = cellList[innerCycle][externalCycle]
 
-            if (shiftCell.value == 0) {
+            if (shiftCell.value == 0.0) {
                 shiftCell.value = currentCell.value
-                currentCell.value = 0
+                currentCell.value = 0.0
                 return@moveUpLeftTemplate ActionMove.EMPTY_MOVE
             } else if (shiftCell.value == currentCell.value) {
                 shiftCell.value *= 2
-                scoreObservable.setValue(scoreObservable.getValue()!! + shiftCell.value)
-                currentCell.value = 0
+                scoreObservable.setValue((scoreObservable.getValue()!! + shiftCell.value).toInt())
+                currentCell.value = 0.0
                 return@moveUpLeftTemplate ActionMove.SUCCESS_MOVE
             } else {
                 return@moveUpLeftTemplate ActionMove.FAILED_MOVE
@@ -106,14 +106,14 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
                 { innerCycle -> (innerCycle + 1) < config.size }) { innerCycle, externalCycle ->
             val shiftCell = cellList[innerCycle + 1][externalCycle]
             val currentCell = cellList[innerCycle][externalCycle]
-            if (shiftCell.value == 0) {
+            if (shiftCell.value == 0.0) {
                 shiftCell.value = currentCell.value
-                currentCell.value = 0
+                currentCell.value = 0.0
                 return@moveDownRightTemplate ActionMove.EMPTY_MOVE
             } else if (shiftCell.value == currentCell.value) {
                 shiftCell.value *= 2
-                scoreObservable.setValue(scoreObservable.getValue()!! + shiftCell.value)
-                currentCell.value = 0
+                scoreObservable.setValue((scoreObservable.getValue()!! + shiftCell.value).toInt())
+                currentCell.value = 0.0
                 return@moveDownRightTemplate ActionMove.SUCCESS_MOVE
             } else {
                 return@moveDownRightTemplate ActionMove.FAILED_MOVE
@@ -126,14 +126,14 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
                 { innerCycle -> (innerCycle + 1) < config.size }) { innerCycle, externalCycle ->
             val shiftCell = cellList[externalCycle][innerCycle + 1]
             val currentCell = cellList[externalCycle][innerCycle]
-            if (shiftCell.value == 0) {
+            if (shiftCell.value == 0.0) {
                 shiftCell.value = currentCell.value
-                currentCell.value = 0
+                currentCell.value = 0.0
                 return@moveDownRightTemplate ActionMove.EMPTY_MOVE
             } else if (currentCell.value == shiftCell.value) {
                 shiftCell.value *= 2
-                scoreObservable.setValue(scoreObservable.getValue()!! + shiftCell.value)
-                currentCell.value = 0
+                scoreObservable.setValue((scoreObservable.getValue()!! + shiftCell.value).toInt())
+                currentCell.value = 0.0
                 return@moveDownRightTemplate ActionMove.SUCCESS_MOVE
             } else {
                 return@moveDownRightTemplate ActionMove.FAILED_MOVE
@@ -146,7 +146,7 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
         var freeCell = 0
         cellList.forEach {
             it.forEach {
-                freeCell += if (it.value == 0) 1 else 0
+                freeCell += if (it.value == 0.0) 1 else 0
             }
         }
         return freeCell
@@ -161,7 +161,7 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
         for (i in 0..(config.size - 1)) {
             for (j in 1..(config.size - 1)) {
                 var temp = j
-                if (startWhilePredicate.invoke(i, j).value != 0) {
+                if (startWhilePredicate.invoke(i, j).value != 0.0) {
                     while (whilePredicate.invoke(temp)) {
                         val actionMove = blockWhile.invoke(temp, i)
                         actionMoveList.add(actionMove)
@@ -185,7 +185,7 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
         for (i in 0..(config.size - 1)) {
             for (j in (config.size - 2) downTo 0) {
                 var temp = j
-                if (startWhilePredicate.invoke(i, j).value != 0) {
+                if (startWhilePredicate.invoke(i, j).value != 0.0) {
                     while (whilePredicate.invoke(temp)) {
                         val actionMove = blockWhile.invoke(temp, i)
                         actionMoveList.add(actionMove)
@@ -214,25 +214,25 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
                     cellHeight.toDouble())
 
             when (cell.value) {
-                0 -> context.fillStyle = "#F3F35F"
-                2 -> context.fillStyle = "#488281"
-                4 -> context.fillStyle = "#C46B3E"
-                8 -> context.fillStyle = "#FF4949"
-                16 -> context.fillStyle = "#9966CC"
-                32 -> context.fillStyle = "#5FACF3"
-                64 -> context.fillStyle = "#8DB600"
-                128 -> context.fillStyle = "#7BA05B"
-                256 -> context.fillStyle = "#00D3B5"
-                512 -> context.fillStyle = "#7FFFD4"
-                1024 -> context.fillStyle = "#4B5320"
-                2048 -> context.fillStyle = "#3B444B"
-                4096 -> context.fillStyle = "#0000FF"
+                0.0 -> context.fillStyle = "#F3F35F"
+                2.0 -> context.fillStyle = "#488281"
+                4.0 -> context.fillStyle = "#C46B3E"
+                8.0 -> context.fillStyle = "#FF4949"
+                16.0 -> context.fillStyle = "#9966CC"
+                32.0 -> context.fillStyle = "#5FACF3"
+                64.0 -> context.fillStyle = "#8DB600"
+                128.0 -> context.fillStyle = "#7BA05B"
+                256.0 -> context.fillStyle = "#00D3B5"
+                512.0 -> context.fillStyle = "#7FFFD4"
+                1024.0 -> context.fillStyle = "#4B5320"
+                2048.0 -> context.fillStyle = "#3B444B"
+                4096.0 -> context.fillStyle = "#0000FF"
                 else -> context.fillStyle = "#007FFF"
             }
 
             context.fill()
 
-            if (cell.value != 0) {
+            if (cell.value != 0.0) {
                 val fontSize = cellWidth / 2f
                 context.font = "${fontSize}px Arial"
                 context.fillStyle = "white"
@@ -243,9 +243,9 @@ object RenderServiceImpl : RenderService, Transformer, ObservableProvider {
     }
 
     private fun createEmptyCell(row: Int, coll: Int): Cell {
-        val x = coll * config.cellWidth + 5 * (coll + 1)
-        val y = row * config.cellHeight + 5 * (row + 1)
-        return Cell(x, y, 0)
+        val x = coll * config.cellWidth.toDouble() + 5 * (coll + 1)
+        val y = row * config.cellHeight.toDouble() + 5 * (row + 1)
+        return Cell(x, y, 0.0)
     }
 
     private fun moveSideEffect(actionMoveList: List<ActionMove>) {
